Adds stubtest for our stubs folder and fixes issues
Adds a weekly workflow run for stubtest
Run stubtest in pre-commit if stubs have been modified
