+++ .github/workflows/stubtest.yml
@@ -1,0 +1,83 @@
+name: Run stubtest on embedded stubs
+
+on:
+  workflow_dispatch:  # allow to run manually
+  schedule:
+    - cron: '0 7 * * 1'  # run once a week on monday morning
+
+jobs:
+  stubtest:
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: Install packages
+      run: |
+        sudo apt update
+        sudo apt install \
+          build-essential \
+          git \
+          gcc \
+          g++ \
+          libcurl4-openssl-dev \
+          libffi-dev \
+          libjpeg-dev \
+          libpq-dev \
+          libxml2-dev \
+          libxslt1-dev \
+          zlib1g-dev \
+          libev-dev \
+          libgnutls28-dev \
+          libkrb5-dev \
+          libpoppler-cpp-dev \
+          pv \
+          libzbar0 \
+          openssl \
+          libssl-dev \
+          xmlsec1 \
+          libxmlsec1-openssl \
+          ghostscript \
+          zip \
+          libev4 \
+          libmagic1 \
+          wget \
+          apparmor \
+          curl \
+          git-core \
+          postgresql \
+          redis \
+          openjdk-8-jre-headless \
+          gcc \
+          libc-dev \
+          fonts-liberation \
+          libappindicator3-1 \
+          libasound2 \
+          libdrm2 \
+          libgbm1 \
+          libnspr4 \
+          libnss3 \
+          libu2f-udev \
+          libvulkan1 \
+          libx11-xcb1 \
+          libxcb-dri3-0 \
+          libxshmfence1 \
+          libxss1 \
+          xdg-utils \
+          --fix-missing
+
+    - name: Checkout repo
+      uses: actions/checkout@v3
+
+    - name: Set up Python
+      uses: actions/setup-python@v4
+      with:
+        python-version: '3.10'
+
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install .[mypy]
+
+    - name: Stubtest
+      run: bash stubtest.sh
+      env:
+        SKIP_VENV: 1
