+++ mypy.sh
@@ -4,8 +4,8 @@ set -euo pipefail
 # change to script directory
 cd $(dirname "$0")
 
-# make sure virtual env is active
-if [ -z "${VIRTUAL_ENV:-}" ]; then
+# make sure virtual env is active unless specifically turned off
+if [[ -z "${VIRTUAL_ENV:-}" && "${SKIP_VENV:-}" -ne "1" ]]; then
     source venv/bin/activate
 fi
