+++ stubs/depot/fields/sqlalchemy.pyi
@@ -6,6 +6,5 @@ from depot.fields.interfaces import FileFilter
 from depot.fields.upload import UploadedFile
 
 class UploadedFileField(types.TypeDecorator):
-    impl = types.Unicode
-    cache_ok: Literal[True]
+    cache_ok: Literal[False]
     def __init__(self, filters: Sequence[FileFilter] = (), upload_type: type[UploadedFile] = ..., upload_storage: str | None = ..., *args: Any, **kw: Any) -> None: ...
